Extract renderer startup migrations
Fixes #154
